Adds AOT compilation support & gulp task Updates Angular to 2.1.1
