+++ package.json
@@ -26,12 +26,13 @@
     "@angular/core": "^2.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^2.0.0",
-    "@angular/compiler": "^2.0.0",
-    "@angular/core": "^2.0.0",
-    "@angular/platform-browser": "^2.0.0",
-    "@angular/platform-browser-dynamic": "^2.0.0",
-    "@angular/router": "3.0.0",
+    "@angular/common": "^2.1.1",
+    "@angular/compiler": "^2.1.1",
+    "@angular/compiler-cli": "^2.1.1",
+    "@angular/core": "^2.1.1",
+    "@angular/platform-browser": "^2.1.1",
+    "@angular/platform-browser-dynamic": "^2.1.1",
+    "@angular/router": "3.1.1",
     "bootstrap": "^3.3.7",
     "chai": "^3.4.1",
     "core-js": "^2.4.1",
@@ -47,14 +48,14 @@
     "gulpclass": "^0.1.0",
     "mocha": "^3.0.2",
     "reflect-metadata": "^0.1.4",
-    "rxjs": "5.0.0-beta.12",
+    "rxjs": "5.0.0-rc.1",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
     "systemjs": "0.19.36",
     "tslint": "^3.14.0",
     "tslint-stylish": "^2.1.0-beta",
-    "typescript": "^1.8.0",
-    "typings": "^1.3.2",
-    "zone.js": "^0.6.21"
+    "typescript": "^2.0.3",
+    "typings": "^1.4.0",
+    "zone.js": "^0.6.26"
   }
 }
