+++ tsconfig-aot.json
@@ -1,0 +1,26 @@
+{
+  "compilerOptions": {
+	"outDir": "build/package",
+    "target": "es5",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "sourceMap": false,
+    "noEmitHelpers": false,
+    "noImplicitAny": true,
+    "declaration": true,
+    "skipLibCheck": true,
+    "stripInternal": true
+  },
+  "exclude": [
+	"build",
+    "node_modules"
+  ],
+  "files": [
+    "./src/index.ts"
+  ],
+  "angularCompilerOptions": {
+    "genDir": "factories"
+  }
+}
